+++ frontend/www/sass/base/_colors.scss
@@ -1,4 +1,5 @@
-$omegaup-white: white ;
+$omegaup-white: white;
+$omegaup-blue: #5588dd;
 $omegaup-pink: #dd5588;
 $omegaup-primary--darker: #466aaf;
 $omegaup-primary--darkest: #3F63AA;
